+++ realm/realm-library/src/main/java/io/realm/BaseRealm.java
@@ -481,8 +481,6 @@ abstract class BaseRealm implements Closeable {
             result = configuration.getSchemaMediator().newInstance(clazz, this, row, schema.getColumnInfo(clazz),
                     false, Collections.<String>emptyList());
         }
-        RealmObjectProxy proxy = (RealmObjectProxy) result;
-        proxy.realmGet$proxyState().setTableVersion$realm();
         return result;
     }
 
@@ -491,8 +489,6 @@ abstract class BaseRealm implements Closeable {
         UncheckedRow row = table.getUncheckedRow(rowIndex);
         E result = configuration.getSchemaMediator().newInstance(clazz, this, row, schema.getColumnInfo(clazz),
                 acceptDefaultValue, excludeFields);
-        RealmObjectProxy proxy = (RealmObjectProxy) result;
-        proxy.realmGet$proxyState().setTableVersion$realm();
         return result;
     }
 
@@ -513,11 +509,6 @@ abstract class BaseRealm implements Closeable {
             result = configuration.getSchemaMediator().newInstance(clazz, this,
                     (rowIndex != Table.NO_MATCH) ? table.getUncheckedRow(rowIndex) : InvalidRow.INSTANCE,
                     schema.getColumnInfo(clazz), false, Collections.<String>emptyList());
-        }
-
-        RealmObjectProxy proxy = (RealmObjectProxy) result;
-        if (rowIndex != Table.NO_MATCH) {
-            proxy.realmGet$proxyState().setTableVersion$realm();
         }
 
         return result;