+++ realm/realm-library/src/main/java/io/realm/internal/SortDescriptor.java
@@ -121,11 +121,6 @@ public class SortDescriptor {
                     "Distinct is not supported on '%s' field '%s' in '%s'.",
                     descriptor.getFieldType().toString(), descriptor.getFieldName(), fieldDescriptions));
         }
-        if (!descriptor.hasSearchIndex()) {
-            throw new IllegalArgumentException(String.format(
-                    "Field '%s' in '%s' must be indexed in order to use it for distinct queries.",
-                    descriptor.getFieldName(), fieldDescriptions));
-        }
     }
 
     // Called by JNI.