+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Breaking Changes
 
+* Updated file format of Realm files. Existing Realm files will automatically be migrated to the new format when they are opened.
+* [ObjectServer] Due to file format changes, Realm Object Server 1.3.0 or later is required.
 * [ObjectServer] Added `onClientResetRequired(SyncSession, ClientResetHandler)` method to the `ErrorHandler` interface (#4080).
 
 ### Enhancements
@@ -12,8 +14,11 @@
   * Linking objects are not yet supported on dynamic objects
   * Migration for linking objects is not yet supported.
   * Backlink verification is incomplete.  Evil code can cause native crashes.
+* [ObjectServer] Added support for Sync Progress Notifications through `SyncSession.addDownloadProgressListener(ProgressMode, ProgressListener)` and `SyncSession.addUploadProgressListener(ProgressMode, ProgressListener)` (#4104).
 * [ObjectServer] In case of a Client Reset, information about the location of the backed up Realm file is now reported through the `ErrorHandler` interface (#4080).
 * [ObjectServer] Authentication URLs now automatically append `/auth` if no other path segment is set (#4370).
+* The listener on `RealmObject` will only be triggered if the object changes (#3894).
+* Added `RealmObjectChangeListener` to get detailed information about `RealmObject` changes.
 
 ### Bug Fixes
 
@@ -22,6 +27,8 @@
 ### Internal
 
 * Using the Object Store's Session and SyncManager.
+* Upgraded to Realm Sync 1.5.0.
+* Upgraded to Realm Core 2.5.1.
 
 
 ## 3.0.1 (YYYY-MM-DD)
@@ -31,6 +38,7 @@
 * Now using Gradle 3.4.1
 * Now `targetSdkVersion` is 25.
 * Listeners on `RealmList` and `RealmResults` will be triggered immediately when the transaction is committed on the same thread (#4245).
+* `RealmQuery.distinct()` can be performed on unindexed fields (#2285).
 
 ### Bug Fixes